Updtaed jwt implementation

+++ src/main/java/controllers/UserController.java
@@ -1,17 +1,16 @@
 package controllers;
 
+import com.ashutosh.springsecurity.exceptions.RecordNotFoundException;
+import com.ashutosh.springsecurity.models.ApiResponse;
 import com.ashutosh.springsecurity.models.User;
 import com.ashutosh.springsecurity.services.UserService;
-import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
-@AllArgsConstructor
 @RequestMapping("api/users")
 public class UserController {
     @Autowired
@@ -19,41 +18,51 @@ public class UserController {
 
     // build create User REST API
     @PostMapping
-    public ResponseEntity<User> saveUser(@RequestBody User user){
+    public ApiResponse saveUser(@RequestBody User user){
         User savedUser = userService.saveUser(user);
-        return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
+        return new ApiResponse(true, savedUser);
     }
 
     // build get user by id REST API
     // /api/users/1
     @GetMapping("{id}")
-    public ResponseEntity<User> getUserById(@PathVariable("id") Long userId){
-        User user = userService.getUserById(userId);
-        return new ResponseEntity<>(user, HttpStatus.OK);
+    public ApiResponse getUserById(@PathVariable("id") Long userId){
+        Optional<User> optionalUser = userService.getUserById(userId);
+        if (optionalUser.isEmpty()) {
+            throw new RecordNotFoundException("User not found with ID: " + userId);
+        } else {
+            return new ApiResponse(true, optionalUser.get());
+        }
     }
 
     // Build Get All Users REST API
     // /api/users
     @GetMapping
-    public ResponseEntity<List<User>> getAllUsers(){
+    public ApiResponse getAllUsers(){
         List<User> users = userService.getAllUsers();
-        return new ResponseEntity<>(users, HttpStatus.OK);
+        return new ApiResponse(true, users);
     }
 
     // Build Update User REST API
     @PutMapping("{id}")
     // /api/users/1
-    public ResponseEntity<User> updateUser(@PathVariable("id") Long userId,
+    public ApiResponse updateUser(@PathVariable("id") Long userId,
                                            @RequestBody User user){
         user.setId(userId);
         User updatedUser = userService.updateUser(user);
-        return new ResponseEntity<>(updatedUser, HttpStatus.OK);
+
+        return new ApiResponse(true, updatedUser);
     }
 
     // Build Delete User REST API
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") Long userId){
-        userService.deleteUser(userId);
-        return new ResponseEntity<>("User successfully deleted!", HttpStatus.OK);
+    public ApiResponse deleteUser(@PathVariable("id") Long userId){
+        Optional<User> optionalUser = userService.getUserById(userId);
+        if (optionalUser.isEmpty()) {
+            throw new RecordNotFoundException("User not found with ID: " + userId);
+        } else {
+            userService.deleteUser(userId);
+            return new ApiResponse(true, "User successfully deleted!");
+        }
     }
 }
